Added a few configs
Most configs don't need to be synced. But for those that I would like to be synced, they will either have their own repository on Github, or be added to the Configs folder of my special Github repository. They will have their own repository if the config is modular and has multiple files, likely without a flat structure, examples would be my BASH config(s) and my Neovim config. If the config is only 1 file, then it will be added to the Configs directory of my special Github repository.